moved incremental cache
