+++ packages/open-next/src/overrides/incrementalCache/s3.ts
@@ -9,7 +9,7 @@ import path from "path";
 
 import { awsLogger } from "../../adapters/logger";
 import { parseNumberFromEnv } from "../../adapters/util";
-import { Extension } from "../next-types";
+import { Extension } from "../../cache/next-types";
 import { IncrementalCache } from "./types";
 
 const {
